+++ src/treaty2/index.ts
@@ -5,7 +5,6 @@ import type { Treaty } from './types'
 import { composePath, isNumericString } from '../treaty/utils'
 import { EdenFetchError } from '../errors'
 import { EdenWS } from './ws'
-import { subscribe } from 'diagnostics_channel'
 
 const method = [
     'get',
@@ -23,11 +22,11 @@ const locals = ['localhost', '127.0.0.1', '0.0.0.0']
 
 const isServer = typeof FileList === 'undefined'
 
-const isISO8601 =
+export const isISO8601 =
     /(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))/
-const isFormalDate =
+export const isFormalDate =
     /(?:Sun|Mon|Tue|Wed|Thu|Fri|Sat)\s(?:Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\s\d{2}\s\d{4}\s\d{2}:\d{2}:\d{2}\sGMT(?:\+|-)\d{4}\s\([^)]+\)/
-const isShortenDate =
+export const isShortenDate =
     /^(?:(?:(?:(?:0?[1-9]|[12][0-9]|3[01])[/\s-](?:0?[1-9]|1[0-2])[/\s-](?:19|20)\d{2})|(?:(?:19|20)\d{2}[/\s-](?:0?[1-9]|1[0-2])[/\s-](?:0?[1-9]|[12][0-9]|3[01]))))(?:\s(?:1[012]|0?[1-9]):[0-5][0-9](?::[0-5][0-9])?(?:\s[AP]M)?)?$/
 
 const isFile = (v: any) => {
@@ -93,6 +92,9 @@ const processHeaders = (
 
     switch (typeof h) {
         case 'function':
+            if (h instanceof Headers)
+                return processHeaders(h, path, options, headers)
+
             const v = h(path, options)
             if (v) return processHeaders(v, path, options, headers)
             return headers
@@ -110,6 +112,67 @@ const processHeaders = (
 
         default:
             return headers
+    }
+}
+
+export async function* streamResponse(response: Response) {
+    const body = response.body
+
+    if (!body) return
+
+    const reader = body.getReader()
+    const decoder = new TextDecoder()
+
+    try {
+        while (true) {
+            const { done, value } = await reader.read()
+            if (done) break
+
+            const data = decoder.decode(value)
+
+            if (isNumericString(data)) {
+                yield +data
+                continue
+            }
+
+            if (data === 'true') {
+                yield true
+                continue
+            }
+
+            if (data === 'false') {
+                yield false
+                continue
+            }
+
+            const start = data.charCodeAt(0)
+            const end = data.charCodeAt(data.length - 1)
+
+            if ((start === 123 && end === 125) || (start === 91 && end === 93))
+                try {
+                    yield JSON.parse(data)
+
+                    continue
+                } catch {}
+
+            // Remove quote from stringified date
+            const temp = data.replace(/"/g, '')
+
+            if (
+                isISO8601.test(temp) ||
+                isFormalDate.test(temp) ||
+                isShortenDate.test(temp)
+            ) {
+                const date = new Date(temp)
+                if (!Number.isNaN(date.getTime())) yield date
+
+                continue
+            }
+
+            yield data
+        }
+    } finally {
+        reader.releaseLock()
     }
 }
 
@@ -279,9 +342,9 @@ const createProxy = (
                             formData.append(key, field as string)
                         }
 
-                    // We don't do this because we need to let the browser set the content type with the correct boundary
-                    // fetchInit.headers['content-type'] = 'multipart/form-data'
-                    fetchInit.body = formData
+                        // We don't do this because we need to let the browser set the content type with the correct boundary
+                        // fetchInit.headers['content-type'] = 'multipart/form-data'
+                        fetchInit.body = formData
                     } else if (typeof body === 'object') {
                         ;(fetchInit.headers as Record<string, string>)[
                             'content-type'
@@ -326,6 +389,7 @@ const createProxy = (
                         new Request(url, fetchInit)
                     ) ?? fetcher!(url, fetchInit))
 
+                    // @ts-ignore
                     let data = null
                     let error = null
 
@@ -353,12 +417,27 @@ const createProxy = (
                         switch (
                             response.headers.get('Content-Type')?.split(';')[0]
                         ) {
+                            case 'text/event-stream':
+                                data = streamResponse(response)
+                                break
+
                             case 'application/json':
                                 data = await response.json()
                                 break
 
                             case 'application/octet-stream':
                                 data = await response.arrayBuffer()
+                                break
+
+                            case 'multipart/form-data':
+                                const temp = await response.formData()
+
+                                data = {}
+                                temp.forEach((value, key) => {
+                                    // @ts-ignore
+                                    data[key] = value
+                                })
+
                                 break
 
                             default: